Changed from GLdouble to GLfloat. Apparently the performance difference is greater on the GPU..? It doesn't really matter but let's be consistent.

+++ src/glRenderer.h
@@ -69,7 +69,7 @@ struct LDGLOverlay
 //
 struct LDGLData
 {
-	GLdouble		rotationMatrix[16];
+	GLfloat			rotationMatrix[16];
 	double			panX[7];
 	double			panY[7];
 	double			zoom[7];
